Rename default package file to public-api
This will be easier to understand when doing an import within the package.
