+++ packages/composite-checkout/webpack.config.js
@@ -1,7 +1,7 @@
 const path = require( 'path' );
 
 module.exports = {
-	entry: './src/index.js',
+	entry: './src/public-api.js',
 	mode: 'development',
 	module: {
 		rules: [
